feat(can): add SocketCAN adapter implementation
Implement a complete SocketCAN adapter for Linux platform, providing
kernel-level CAN communication interface support.

Features:
- Full SocketCAN integration using socketcan crate (v3.5)
- Support for standard and extended CAN frames
- Hardware timestamp support with automatic fallback to software timestamps
- Automatic error frame filtering
- Non-blocking receive with configurable timeout
- Integration with existing CanAdapter trait

Changes:
- Add src/can/socketcan/mod.rs with complete SocketCAN implementation
- Update Cargo.toml with socketcan dependency
- Integrate SocketCAN adapter into can module exports
- Update robot builder and pipeline to support SocketCAN
- Add comprehensive documentation including implementation plans,
  hardware timestamp implementation plan, and TODO lists
- Add timestamp verification example
- Update tests to work with new adapter

Documentation:
- Add implementation_plan.md with design overview
- Add hardware_timestamp_implementation_plan.md with detailed HW timestamp design
- Add HARDWARE_TIMESTAMP_TODO.md with tracking items
- Add TODO_LIST.md for remaining tasks
- Add survey.md with research findings

This implementation enables native Linux CAN support through the SocketCAN
subsystem, offering better performance compared to userspace drivers like
GS-USB, while maintaining compatibility with the existing CanAdapter interface.

+++ src/can/gs_usb/mod.rs
@@ -329,7 +329,7 @@ impl CanAdapter for GsUsbCanAdapter {
                     data: gs_frame.data,
                     len: gs_frame.can_dlc.min(8),
                     is_extended: (gs_frame.can_id & CAN_EFF_FLAG) != 0,
-                    timestamp_us: gs_frame.timestamp_us, // 保留硬件时间戳
+                    timestamp_us: gs_frame.timestamp_us as u64, // 保留硬件时间戳（GS-USB 使用 u32，转换为 u64）
                 };
 
                 self.rx_queue.push_back(frame);
